requirements.txt
astroid==2.2.5
colorama==0.4.1
isort==4.3.18
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==2.3.1
six==1.12.0
typed-ast==1.3.5
wrapt==1.11.1
biopython==1.73